+++ PecTec.Riot.LoL/Models/Champions/Specifics/Stats.cs
@@ -2,25 +2,25 @@
 {
     public class Stats
     {
-        public int Hp { get; set; }
-        public int Hpperlevel { get; set; }
-        public int Mp { get; set; }
-        public int Mpperlevel { get; set; }
-        public int Movespeed { get; set; }
-        public int Armor { get; set; }
-        public double Armorperlevel { get; set; }
-        public int Spellblock { get; set; }
-        public double Spellblockperlevel { get; set; }
-        public int Attackrange { get; set; }
-        public int Hpregen { get; set; }
-        public int Hpregenperlevel { get; set; }
-        public int Mpregen { get; set; }
-        public int Mpregenperlevel { get; set; }
-        public int Crit { get; set; }
-        public int Critperlevel { get; set; }
-        public int Attackdamage { get; set; }
-        public int Attackdamageperlevel { get; set; }
-        public double Attackspeedperlevel { get; set; }
-        public double Attackspeed { get; set; }
+        public decimal Hp { get; set; }
+        public decimal Hpperlevel { get; set; }
+        public decimal Mp { get; set; }
+        public decimal Mpperlevel { get; set; }
+        public decimal Movespeed { get; set; }
+        public decimal Armor { get; set; }
+        public decimal Armorperlevel { get; set; }
+        public decimal Spellblock { get; set; }
+        public decimal Spellblockperlevel { get; set; }
+        public decimal Attackrange { get; set; }
+        public decimal Hpregen { get; set; }
+        public decimal Hpregenperlevel { get; set; }
+        public decimal Mpregen { get; set; }
+        public decimal Mpregenperlevel { get; set; }
+        public decimal Crit { get; set; }
+        public decimal Critperlevel { get; set; }
+        public decimal Attackdamage { get; set; }
+        public decimal Attackdamageperlevel { get; set; }
+        public decimal Attackspeedperlevel { get; set; }
+        public decimal Attackspeed { get; set; }
     }
 }
